+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## Unreleased
+### Added
+- [WebSocket Trading](https://bybit-exchange.github.io/docs/v5/websocket/trade/guideline) support
+
+
 ## [5.7.0] - 2024-04-11
 ### Added
 - [Demo Trading](https://bybit-exchange.github.io/docs/v5/demo) support